move score file to /entities and move interfaces from WizardState
